Bugfix
Reported by Romain Diss on context-mailing-list.

style and color parameters were not working.

This is because context parses the parameter of `\useMPgraphic` to determine
the MP instance. The code was using

    \useMPgraphic{\visualcounterparameter\c!mp}

`\useMPgraphic` does not expand its arguments, so the proper MP instance was
not initialized. Changed the above line to

    \expanded{\useMPgraphic{\visualcounterparameter\c!mp}}

and now style and color parameters work.